Remove old h5 tests
